实现未识别列表响应式更新

+++ src-tauri/src/service/unrecognized_videos.rs
@@ -1,6 +1,10 @@
 use crate::{
-    data::{pending_videos, unrecognized_videos::VideoData},
+    data::{
+        pending_videos,
+        unrecognized_videos::{get_all, VideoData},
+    },
     model::setting::Setting,
+    utils::tauri::send_event,
 };
 use std::path::Path;
 
@@ -10,11 +14,13 @@ pub fn delete<P: AsRef<Path>>(path: P) -> Result<(), UnrecognizedVideosServiceEr
     // 如果存在文件，同时删除文件
     if path.exists()
         && path.is_file()
+        && !path.is_symlink()
         && pending_videos::get(&path).is_none()
         && path.starts_with(Setting::get_pending_path().as_path())
     {
         std::fs::remove_file(path)?;
     }
+    send_event("main", "unrecognized_videos_list", get_all());
     Ok(())
 }
 
@@ -37,6 +43,7 @@ pub async fn insert<P: AsRef<Path>>(
         }
         VideoData::Undefined => {
             crate::data::unrecognized_videos::insert(path, video_data)?;
+            send_event("main", "unrecognized_videos_list", get_all());
         }
     };
